+++ src/contour/TerminalSession.h
@@ -135,9 +135,12 @@ class TerminalSession: public QObject, public terminal::Terminal::Events
     bool operator()(actions::DecreaseFontSize);
     bool operator()(actions::DecreaseOpacity);
     bool operator()(actions::FollowHyperlink);
+    bool operator()(actions::FocusNextSearchMatch);
+    bool operator()(actions::FocusPreviousSearchMatch);
     bool operator()(actions::IncreaseFontSize);
     bool operator()(actions::IncreaseOpacity);
     bool operator()(actions::NewTerminal const&);
+    bool operator()(actions::NoSearchHighlight);
     bool operator()(actions::OpenConfiguration);
     bool operator()(actions::OpenFileManager);
     bool operator()(actions::PasteClipboard);
@@ -157,6 +160,7 @@ class TerminalSession: public QObject, public terminal::Terminal::Events
     bool operator()(actions::ScrollToBottom);
     bool operator()(actions::ScrollToTop);
     bool operator()(actions::ScrollUp);
+    bool operator()(actions::SearchReverse);
     bool operator()(actions::SendChars const& _event);
     bool operator()(actions::ToggleAllKeyMaps);
     bool operator()(actions::ToggleFullscreen);