+++ src/terminal/Line.h
@@ -21,6 +21,8 @@
 #include <crispy/Comparison.h>
 #include <crispy/assert.h>
 
+#include <unicode/convert.h>
+
 #include <gsl/span>
 #include <gsl/span_ext>
 
@@ -224,7 +226,7 @@ class Line
         return inflatedBuffer().at(unbox<size_t>(column)).empty();
     }
 
-    [[nodiscard]] uint8_t cellWithAt(ColumnOffset column) const noexcept
+    [[nodiscard]] uint8_t cellWidthAt(ColumnOffset column) const noexcept
     {
         if (isTrivialBuffer())
         {
@@ -345,6 +347,116 @@ class Line
     {
         storage_ =
             TrivialBuffer { size(), textAttributes, fillAttributes, hyperlink, columnsUsed, std::move(text) };
+    }
+
+    // Tests if the given text can be matched in this line at the exact given start column.
+    [[nodiscard]] bool matchTextAt(std::u32string_view text, ColumnOffset startColumn) const noexcept
+    {
+        if (isTrivialBuffer())
+        {
+            auto const u8Text = unicode::convert_to<char>(text);
+            TrivialBuffer const& buffer = trivialBuffer();
+            if (!buffer.usedColumns)
+                return false;
+            auto const column = std::min(startColumn, boxed_cast<ColumnOffset>(buffer.usedColumns - 1));
+            auto const resultIndex = buffer.text.view()
+                                         .substr(unbox<size_t>(column))
+                                         .find(std::string_view(u8Text), unbox<size_t>(column));
+            return resultIndex == 0;
+        }
+        else
+        {
+            auto const u8Text = unicode::convert_to<char>(text);
+            InflatedBuffer const& cells = inflatedBuffer();
+            auto const baseColumn = unbox<size_t>(startColumn);
+            size_t i = 0;
+            while (i < text.size())
+            {
+                if (!beginsWith(text.substr(i), cells[baseColumn + i]))
+                    return false;
+                ++i;
+            }
+            return i == text.size();
+        }
+    }
+
+    // Searches for the given text in this line starting at a specific column.
+    // Returns either the location of the beginning of found search term or the same input start column if
+    // not.
+    [[nodiscard]] std::optional<ColumnOffset> search(std::u32string_view text,
+                                                     ColumnOffset startColumn) const noexcept
+    {
+        if (isTrivialBuffer())
+        {
+            auto const u8Text = unicode::convert_to<char>(text);
+            TrivialBuffer const& buffer = trivialBuffer();
+            if (!buffer.usedColumns)
+                return std::nullopt;
+            auto const column = std::min(startColumn, boxed_cast<ColumnOffset>(buffer.usedColumns - 1));
+            auto const resultIndex = buffer.text.view().find(std::string_view(u8Text), unbox<size_t>(column));
+            if (resultIndex != std::string_view::npos)
+                return ColumnOffset::cast_from(resultIndex);
+            else
+                return std::nullopt; // Not found, so stay with initial column as result.
+        }
+        else
+        {
+            InflatedBuffer const& buffer = inflatedBuffer();
+            if (buffer.size() < text.size())
+                return std::nullopt; // not found: line is smaller than search term
+
+            auto baseColumn = std::min(startColumn, ColumnOffset::cast_from(buffer.size() - text.size()));
+            auto rightMostSearchPosition =
+                std::max(startColumn, ColumnOffset::cast_from(buffer.size() - text.size()));
+            while (baseColumn < rightMostSearchPosition)
+            {
+                if (matchTextAt(text, baseColumn))
+                    return baseColumn;
+                baseColumn++;
+            }
+
+            return std::nullopt; // Not found, so stay with initial column as result.
+        }
+    }
+
+    // Searches for the given input text in this line starting at a specific column from right to left.
+    //
+    // The column offset to the beginning of the match will be returned if found.
+    // If no match could be found, the input column will be returned instead.
+    [[nodiscard]] ColumnOffset searchReverse(std::u32string_view text,
+                                             ColumnOffset startColumn) const noexcept
+    {
+        if (isTrivialBuffer())
+        {
+            auto const u8Text = unicode::convert_to<char>(text);
+            TrivialBuffer const& buffer = trivialBuffer();
+            if (!buffer.usedColumns)
+                return startColumn;
+            auto const column = std::min(startColumn, boxed_cast<ColumnOffset>(buffer.usedColumns - 1));
+            auto const resultIndex =
+                buffer.text.view().rfind(std::string_view(u8Text), unbox<size_t>(column));
+            if (resultIndex != std::string_view::npos)
+                return ColumnOffset::cast_from(resultIndex);
+            else
+                return startColumn; // Not found, so stay with initial column as result.
+        }
+        else
+        {
+            InflatedBuffer const& buffer = inflatedBuffer();
+            if (buffer.size() < text.size())
+                return startColumn; // not found: line is smaller than search term
+
+            // reverse search from right@column to left until match is complete.
+            auto baseColumn = std::min(startColumn, ColumnOffset::cast_from(buffer.size() - text.size()));
+            while (baseColumn >= ColumnOffset(0))
+            {
+                if (matchTextAt(text, baseColumn))
+                    return baseColumn;
+                baseColumn--;
+            }
+
+            return startColumn; // Not found, so stay with initial column as result.
+        }
     }
 
   private: